Set defails for Actions, added defaults for placeholds in Progreesion
 * Also made variance a long, because it gets too big, too fast to be
interesting otherwise.

+++ src/main/java/ca/concordia/moloch/tileentity/moloch/action/ActionOutputMapper.java
@@ -36,7 +36,7 @@ public class ActionOutputMapper {
 		nbt.putInt("doCountTotal", a.getDoCountTotal());
 		nbt.putInt("doCountRemaining", a.getDoCountRemaining());
 		nbt.putInt("interval", a.getInterval());
-		nbt.putInt("variance", a.getVariance());
+		nbt.putLong("variance", a.getVariance());
 		nbt.putLong("lastRun", a.getLastRun());
 		nbt.putBoolean("active", a.isActive());
